refactor(evals): split runner into lib and qbds plugin
